Created UI for group creation, group joining and list of group page
This partially completes the JIRA ticket "KAN-20". All of the new components have been designed such that it can be navigated in the website itself without the need to directly type in a specific URL.

I don't know what the UI should look like for group management, so I excluded that out.